requirements.txt
apache-airflow-providers-celery
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
pandas
requests
openpyxl
python-dotenv
dbt-core==1.8.0
dbt-snowflake==1.8.0
sqlalchemy
snowflake-connector-python
snowflake-sqlalchemy
boto3
minio